+++ core/dbt/deprecations.py
@@ -57,22 +57,6 @@ class DispatchPackagesDeprecation(DBTDeprecation):
     '''
 
 
-class ModelsKeyNonModelDeprecation(DBTDeprecation):
-    _name = 'models-key-mismatch'
-
-    _description = '''\
-    "{node.name}" is a {node.resource_type} node, but it is specified in
-    the {patch.yaml_key} section of {patch.original_file_path}.
-
-
-
-    To fix this warning, place the `{node.name}` specification under
-    the {expected_key} key instead.
-
-    This warning will become an error in a future release.
-    '''
-
-
 class PackageRedirectDeprecation(DBTDeprecation):
     _name = 'package-redirect'
     _description = '''\
@@ -120,7 +104,6 @@ active_deprecations: Set[str] = set()
 
 deprecations_list: List[DBTDeprecation] = [
     DispatchPackagesDeprecation(),
-    ModelsKeyNonModelDeprecation(),
     PackageRedirectDeprecation()
 ]