+++ core/dbt/exceptions.py
@@ -5,7 +5,7 @@ from typing import NoReturn, Optional, Mapping, Any
 from dbt.logger import GLOBAL_LOGGER as logger
 from dbt.node_types import NodeType
 from dbt import flags
-from dbt.ui import line_wrap_message
+from dbt.ui import line_wrap_message, warning_tag
 
 import dbt.dataclass_schema
 
@@ -924,22 +924,17 @@ def raise_unrecognized_credentials_type(typename, supported_types):
     )
 
 
-def raise_invalid_patch(
-    node, patch_section: str, patch_path: str,
-) -> NoReturn:
+def warn_invalid_patch(patch, resource_type):
     msg = line_wrap_message(
         f'''\
-        '{node.name}' is a {node.resource_type} node, but it is
-        specified in the {patch_section} section of
-        {patch_path}.
-
-
-
-        To fix this error, place the `{node.name}`
-        specification under the {node.resource_type.pluralize()} key instead.
+        '{patch.name}' is a {resource_type} node, but it is
+        specified in the {patch.yaml_key} section of
+        {patch.original_file_path}.
+        To fix this error, place the `{patch.name}`
+        specification under the {resource_type.pluralize()} key instead.
         '''
     )
-    raise_compiler_error(msg, node)
+    warn_or_error(msg, log_fmt=warning_tag('{}'))
 
 
 def raise_not_implemented(msg):